refactor: [EL-4914] update import paths for tour target constants
