+++ src/[fsd]/features/interactive-tours/lib/constants/agentTourTargets.constants.js
