+++ src/[fsd]/features/interactive-tours/lib/constants/chatTourTargets.constants.js
