+++ src/[fsd]/features/interactive-tours/lib/constants/index.js
@@ -1,6 +1,6 @@
 export { CARD_WIDTH_PX } from './interactiveTour.constants';
 export { AGENT_TOUR_COMPLETION, AGENT_TOUR_ID, agentTourSteps } from './agentTour.constants';
-export { AGENT_TOUR_TARGET_IDS, AGENT_TOUR_TARGETS } from './agentTourTargets.js';
+export { AGENT_TOUR_TARGET_IDS, AGENT_TOUR_TARGETS } from './agentTourTargets.constants.js';
 export {
   CONVERSATION_TOUR_COMPLETION,
   CONVERSATION_TOUR_ID,
@@ -12,6 +12,6 @@ export {
   firstEliteaTourSteps,
 } from './firstEliteaTour.constants';
 export { CHAT_TOUR_COMPLETION, CHAT_TOUR_ID, chatTourSteps } from './chatTour.constants';
-export { CHAT_TOUR_TARGET_IDS, CHAT_TOUR_TARGETS } from './chatTourTargets.js';
+export { CHAT_TOUR_TARGET_IDS, CHAT_TOUR_TARGETS } from './chatTourTargets.constants.js';
 export { SIDEBAR_TOUR_COMPLETION, SIDEBAR_TOUR_ID, sidebarTourSteps } from './sidebarTour.constants';
 export { SIDEBAR_TOUR_TARGET_IDS, SIDEBAR_TOUR_TARGETS } from './sidebarTourTargets.constants.js';
